wp: test is its own module
